+++ lib/framework/ESPUtils.h
@@ -7,7 +7,7 @@ class ESPUtils {
   public:
     static String defaultDeviceValue(String prefix = "") {
 #ifdef ESP32
-        return prefix + String((unsigned long)ESP.getEfuseMac(), HEX);
+        return prefix + String((uint32_t)ESP.getEfuseMac(), HEX);
 #elif defined(ESP8266)
         return prefix + String(ESP.getChipId(), HEX);
 #endif
